Remove test import * statements
These were convenience imports in the __init__.py, but the behavior of python2 and python3 test runners differ in that python3 will actually run tests imported that way,
respecting that tests/__init__.py is actually a test package and even if the modules within are not prefixed with `test_`, tests exist. Python2 does not do this and requires
specifying test files or requires test files be prefixed. Renaming the files to match the common case makes python2 and python3 work without having to specify test files, but
a side effect of that is python3 handles both cases and runs tests twice